Rename `src` to `vim_python_domain_knowledge` to make it compatible
