+++ plugin/vim_python_domain_knowledge/common/data_structures.py
