+++ plugin/vim_python_domain_knowledge/settings.py
@@ -1,4 +1,4 @@
-from src.common.vim import Vim
+from vim_python_domain_knowledge.common.vim import Vim
 
 CURRENT_DIRECTORY = Vim.eval("getcwd()")
 KNOWLEDGE_DIRECTORY = f'{CURRENT_DIRECTORY}/.vim_domain_knowledge/'
