+++ plugin/vim_python_domain_knowledge/setup.cfg
